Refactor: Move the SwapMove classes to the correct package
