+++ src/main/java/com/suse/matcher/solver/MatchSwapMoveIterator.java
@@ -13,7 +13,7 @@
  * in this software or its documentation.
  */
 
-package com.suse.matcher;
+package com.suse.matcher.solver;
 
 import static java.util.stream.Collectors.groupingBy;
 import static java.util.stream.Collectors.toList;
